moved high score logic out of app delete save data on game over
